requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.47.0
MarkupSafe==3.0.3
more-itertools==11.0.1
mpmath==1.3.0
networkx==3.6.1
numba==0.65.0
numpy==2.4.4
openai-whisper==20250625
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.24
regex==2026.4.4
requests==2.33.1
setuptools==81.0.0
starlette==1.0.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.11.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
